Start using local token
